always pass A as float64 array in tests

+++ tests/test_doolittle_factorization.py
@@ -8,14 +8,14 @@ from tests.helper import check_frobenius_norm_lu
 @pytest.mark.parametrize(
     "A, L_expected, U_expected",
     [
-        (np.array([[2]]), np.array([[1]]), np.array([[2]])),
+        (np.array([[2]], dtype=np.float64), np.array([[1]]), np.array([[2]])),
         (
-            np.array([[1, 2], [3, 4]]),
+            np.array([[1, 2], [3, 4]], dtype=np.float64),
             np.array([[1, 0], [3, 1]]),
             np.array([[1, 2], [0, -2]]),
         ),
         (
-            np.array([[2, -1, -2], [-4, 6, 3], [-4, -2, 8]]),
+            np.array([[2, -1, -2], [-4, 6, 3], [-4, -2, 8]], dtype=np.float64),
             np.array([[1, 0, 0], [-2, 1, 0], [-2, -1, 1]]),
             np.array([[2, -1, -2], [0, 4, -1], [0, 0, 3]]),
         ),
@@ -31,14 +31,14 @@ def test_doolittle_factorization_sequential(A, L_expected, U_expected):
 @pytest.mark.parametrize(
     "A, L_expected, U_expected",
     [
-        (np.array([[2]]), np.array([[1]]), np.array([[2]])),
+        (np.array([[2]], dtype=np.float64), np.array([[1]]), np.array([[2]])),
         (
-            np.array([[1, 2], [3, 4]]),
+            np.array([[1, 2], [3, 4]], dtype=np.float64),
             np.array([[1, 0], [3, 1]]),
             np.array([[1, 2], [0, -2]]),
         ),
         (
-            np.array([[2, -1, -2], [-4, 6, 3], [-4, -2, 8]]),
+            np.array([[2, -1, -2], [-4, 6, 3], [-4, -2, 8]], dtype=np.float64),
             np.array([[1, 0, 0], [-2, 1, 0], [-2, -1, 1]]),
             np.array([[2, -1, -2], [0, 4, -1], [0, 0, 3]]),
         ),
@@ -55,14 +55,14 @@ def test_doolittle_factorization_parallel(A, L_expected, U_expected):
 @pytest.mark.parametrize(
     "A, L_expected, U_expected",
     [
-        (np.array([[2]]), np.array([[1]]), np.array([[2]])),
+        (np.array([[2]], dtype=np.float64), np.array([[1]]), np.array([[2]])),
         (
-            np.array([[1, 2], [3, 4]]),
+            np.array([[1, 2], [3, 4]], dtype=np.float64),
             np.array([[1, 0], [3, 1]]),
             np.array([[1, 2], [0, -2]]),
         ),
         (
-            np.array([[2, -1, -2], [-4, 6, 3], [-4, -2, 8]]),
+            np.array([[2, -1, -2], [-4, 6, 3], [-4, -2, 8]], dtype=np.float64),
             np.array([[1, 0, 0], [-2, 1, 0], [-2, -1, 1]]),
             np.array([[2, -1, -2], [0, 4, -1], [0, 0, 3]]),
         ),
@@ -79,14 +79,14 @@ def test_doolittle_factorization_parallel_futures(A, L_expected, U_expected, n_t
 @pytest.mark.parametrize(
     "A, L_expected, U_expected",
     [
-        (np.array([[2]]), np.array([[1]]), np.array([[2]])),
+        (np.array([[2]], dtype=np.float64), np.array([[1]]), np.array([[2]])),
         (
-            np.array([[1, 2], [3, 4]]),
+            np.array([[1, 2], [3, 4]], dtype=np.float64),
             np.array([[1, 0], [3, 1]]),
             np.array([[1, 2], [0, -2]]),
         ),
         (
-            np.array([[2, -1, -2], [-4, 6, 3], [-4, -2, 8]]),
+            np.array([[2, -1, -2], [-4, 6, 3], [-4, -2, 8]], dtype=np.float64),
             np.array([[1, 0, 0], [-2, 1, 0], [-2, -1, 1]]),
             np.array([[2, -1, -2], [0, 4, -1], [0, 0, 3]]),
         ),
